chatbot/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
langchain==0.3.7
langchain-groq
langchain-google-genai==2.0.8
langchain-community==0.3.7
supabase==2.4.6
httpx[http2]>=0.24,<0.28
python-dotenv==1.0.1
pymysql==1.1.0
